Fix race condition which permanently pauses sync
The race condition occurs when the block queue is full so syncing is
paused, and the function which detects when the block queue has room
(BlockQueue::drain) does the first of the two-part check when a verifier
thread has temporarily removed block data from the verification queues.
As such, the detection fails and drain removes more blocks, which means
that the block queue enters a permanently not full state and syncing
remains paused indefinitely.

This is fixed by also performing the "room available" detection check in
the body of the block queue verifier threads.